requirements.txt
streamlit
pandas==2.1.0
plotly==5.15.0
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
ofxparse==0.21
langchain-openai==0.0.8
langchain-groq==0.0.5
langchain-core==0.1.0
numpy==1.24.3
openpyxl==3.1.2
